move relic out of explore
