+++ src/main/java/ti4/commands/explore/ExploreCommand.java
@@ -1,17 +1,19 @@
 package ti4.commands.explore;
 
+import java.util.Collection;
+import java.util.HashSet;
+import java.util.Objects;
+
 import net.dv8tion.jda.api.events.interaction.command.SlashCommandInteractionEvent;
 import net.dv8tion.jda.api.interactions.commands.build.Commands;
 import net.dv8tion.jda.api.requests.restaction.CommandListUpdateAction;
 import ti4.commands.Command;
+import ti4.commands.relic.PurgeFragments;
+import ti4.commands.relic.SendFragments;
 import ti4.helpers.Constants;
 import ti4.map.Game;
 import ti4.map.GameManager;
 import ti4.map.GameSaveLoadManager;
-
-import java.util.Collection;
-import java.util.HashSet;
-import java.util.Objects;
 
 public class ExploreCommand implements Command {
 
@@ -57,21 +59,8 @@ public class ExploreCommand implements Command {
         subcommands.add(new ExpPlanet());
         subcommands.add(new ExpReset());
         subcommands.add(new ExpFrontier());
-        subcommands.add(new SendFragments());
         subcommands.add(new UseExplore());
-        subcommands.add(new PurgeFragments());
-        subcommands.add(new DrawRelic());
-        subcommands.add(new PurgeRelic());
-        subcommands.add(new ShuffleRelicBack());
-        subcommands.add(new ExhaustRelic());
-        subcommands.add(new RefreshRelic());
-        subcommands.add(new DrawSpecificRelic());
-        subcommands.add(new ShowRemainingRelics());
-        subcommands.add(new AddRelicBackIntoDeck());
-        subcommands.add(new RelicInfo());
-        subcommands.add(new RelicLookAtTop());
         subcommands.add(new ExploreLookAtTop());
-        subcommands.add(new RelicSend());
 
         return subcommands;
     }
