+++ src/main/java/ti4/commands/leaders/HeroPlay.java
@@ -13,8 +13,8 @@ import net.dv8tion.jda.api.interactions.components.buttons.Button;
 import ti4.buttons.Buttons;
 import ti4.commands.agenda.DrawAgenda;
 import ti4.commands.cardsac.ACInfo;
-import ti4.commands.explore.DrawRelic;
 import ti4.commands.planet.PlanetRefresh;
+import ti4.commands.relic.DrawRelic;
 import ti4.commands.special.KeleresHeroMentak;
 import ti4.commands.special.RiseOfMessiah;
 import ti4.commands.status.ListTurnOrder;
