+++ src/main/java/ti4/commands/relic/ExhaustRelic.java
@@ -1,4 +1,4 @@
-package ti4.commands.explore;
+package ti4.commands.relic;
 
 import net.dv8tion.jda.api.events.interaction.command.SlashCommandInteractionEvent;
 import net.dv8tion.jda.api.interactions.commands.OptionMapping;
@@ -7,39 +7,47 @@ import net.dv8tion.jda.api.interactions.commands.build.OptionData;
 import ti4.generator.Mapper;
 import ti4.helpers.Constants;
 import ti4.helpers.Emojis;
+import ti4.helpers.Helper;
+import ti4.map.Game;
 import ti4.map.Player;
 import ti4.message.MessageHelper;
 
-public class ExhaustRelic extends GenericRelicAction {
+public class ExhaustRelic extends RelicSubcommandData {
 
     public ExhaustRelic() {
-        super(Constants.RELIC_EXHAUST, "Exhaust a relic", true);
+        super(Constants.RELIC_EXHAUST, "Exhaust a Relic");
         addOptions(new OptionData(OptionType.STRING, Constants.RELIC, "Relic to exhaust").setAutoComplete(true).setRequired(true));
         addOptions(new OptionData(OptionType.STRING, Constants.FACTION_COLOR, "Faction or Color").setAutoComplete(true));
     }
 
 	public ExhaustRelic(String relicRefresh, String refresh_a_relic) {
-		super(relicRefresh, refresh_a_relic, true);
+		super(relicRefresh, refresh_a_relic);
 	}
 
 	@Override
-    public void doAction(Player player, SlashCommandInteractionEvent event) {
+    public void execute(SlashCommandInteractionEvent event) {
+        Game game = getActiveGame();
+        Player player = game.getPlayer(getUser().getId());
+        player = Helper.getGamePlayer(game, player, event, null);
+        player = Helper.getPlayer(game, player, event);
+        if (player == null) {
+            MessageHelper.sendMessageToEventChannel(event, "Player could not be found");
+            return;
+        }
+
         OptionMapping option = event.getOption(Constants.RELIC);
         if (option == null) {
             MessageHelper.sendMessageToEventChannel(event, "Specify relic");
             return;
         }
+
         String relicId = option.getAsString();
         if (player.hasRelic(relicId)) {
-			subAction(player, event, relicId);
+            player.addExhaustedRelic(relicId);
+            String relicName = Mapper.getRelic(relicId).getName();
+            MessageHelper.sendMessageToEventChannel(event, "Exhausted " + Emojis.Relic + " relic: " + relicName);
 		} else {
             MessageHelper.sendMessageToEventChannel(event, "Invalid relic or player does not have specified relic");
         }
     }
-
-	protected void subAction(Player player, SlashCommandInteractionEvent event, String relicId) {
-		player.addExhaustedRelic(relicId);
-		String relicName = Mapper.getRelic(relicId).getName();
-		MessageHelper.sendMessageToEventChannel(event, "Exhausted " + Emojis.Relic + " relic: " + relicName);
-	}
 }
