+++ src/main/java/ti4/commands/relic/PurgeRelic.java
@@ -1,0 +1,45 @@
+package ti4.commands.relic;
+
+import net.dv8tion.jda.api.events.interaction.command.SlashCommandInteractionEvent;
+import net.dv8tion.jda.api.interactions.commands.OptionMapping;
+import net.dv8tion.jda.api.interactions.commands.OptionType;
+import net.dv8tion.jda.api.interactions.commands.build.OptionData;
+import ti4.generator.Mapper;
+import ti4.helpers.Constants;
+import ti4.helpers.Emojis;
+import ti4.helpers.Helper;
+import ti4.map.Game;
+import ti4.map.Player;
+import ti4.message.MessageHelper;
+import ti4.model.RelicModel;
+
+public class PurgeRelic extends RelicSubcommandData {
+
+    public PurgeRelic() {
+        super(Constants.RELIC_PURGE, "Purge a relic");
+        addOptions(new OptionData(OptionType.STRING, Constants.RELIC, "Relic to purge").setAutoComplete(true).setRequired(true));
+        addOptions(new OptionData(OptionType.STRING, Constants.FACTION_COLOR, "Faction or Color").setAutoComplete(true));
+    }
+
+    @Override
+    public void execute(SlashCommandInteractionEvent event) {
+        Game game = getActiveGame();
+        Player player = game.getPlayer(getUser().getId());
+        player = Helper.getGamePlayer(game, player, event, null);
+        player = Helper.getPlayer(game, player, event);
+        if (player == null) {
+            MessageHelper.sendMessageToEventChannel(event, "Player could not be found");
+            return;
+        }
+        String relicId = event.getOption(Constants.RELIC, null, OptionMapping::getAsString);
+        if (relicId == null || !player.hasRelic(relicId)) {
+            MessageHelper.sendMessageToEventChannel(event, "Invalid relic or player does not have specified relic: " + relicId);
+            return;
+        }
+        player.removeRelic(relicId);
+        player.removeExhaustedRelic(relicId);
+        RelicModel relicData = Mapper.getRelic(relicId);
+        MessageHelper.sendMessageToEventChannel(event, player.getRepresentation() + " purged relic Relic:\n" + Emojis.Relic + " __**" + relicData.getName() + "**__\n> " + relicData.getText() + "\n");
+        RelicInfo.sendRelicInfo(getActiveGame(), player, event);
+    }
+}
