+++ src/main/java/ti4/commands/relic/RelicLookAtTop.java
@@ -1,0 +1,38 @@
+package ti4.commands.relic;
+
+import java.util.List;
+
+import net.dv8tion.jda.api.events.interaction.command.SlashCommandInteractionEvent;
+import ti4.generator.Mapper;
+import ti4.helpers.Constants;
+import ti4.helpers.Helper;
+import ti4.map.Game;
+import ti4.map.Player;
+import ti4.message.MessageHelper;
+import ti4.model.RelicModel;
+
+public class RelicLookAtTop extends RelicSubcommandData {
+    public RelicLookAtTop() {
+        super(Constants.RELIC_LOOK_AT_TOP, "Look at the top of the relic deck. Sends to Cards Info thread.");
+    }
+
+    public void execute(SlashCommandInteractionEvent event) {
+        Game game = getActiveGame();
+        Player player = game.getPlayer(getUser().getId());
+        player = Helper.getGamePlayer(game, player, event, null);
+        player = Helper.getPlayer(game, player, event);
+        if (player == null) {
+            MessageHelper.sendMessageToEventChannel(event, "Player could not be found");
+            return;
+        }
+        List<String> relicDeck = game.getAllRelics();
+        if (relicDeck.isEmpty()) {
+            MessageHelper.sendMessageToEventChannel(event, "Relic deck is empty");
+            return;
+        }
+        String relicID = relicDeck.get(0);
+        RelicModel relicModel = Mapper.getRelic(relicID);
+        String sb = "**Relic - Look at Top**\n" + player.getRepresentation() + "\n" + relicModel.getSimpleRepresentation();
+        MessageHelper.sendMessageToPlayerCardsInfoThread(player, game, sb);
+    }
+}
