+++ src/main/java/ti4/commands/relic/RelicSubcommandData.java
@@ -1,0 +1,41 @@
+package ti4.commands.relic;
+
+import net.dv8tion.jda.api.entities.User;
+import net.dv8tion.jda.api.events.interaction.command.SlashCommandInteractionEvent;
+import net.dv8tion.jda.api.interactions.commands.OptionType;
+import net.dv8tion.jda.api.interactions.commands.build.OptionData;
+import net.dv8tion.jda.api.interactions.commands.build.SubcommandData;
+import ti4.helpers.Constants;
+import ti4.map.Game;
+import ti4.map.GameManager;
+
+abstract class RelicSubcommandData extends SubcommandData {
+
+    private Game game;
+    private User user;
+    protected final OptionData typeOption = new OptionData(OptionType.STRING, Constants.TRAIT, "Cultural, Industrial, Hazardous, or Frontier.").setAutoComplete(true);
+
+    public String getActionID() {
+        return getName();
+    }
+
+    public RelicSubcommandData(String name, String description) {
+        super(name, description);
+    }
+
+    public Game getActiveGame() {
+        return game;
+    }
+
+    public User getUser() {
+        return user;
+    }
+
+    public abstract void execute(SlashCommandInteractionEvent event);
+
+    public void preExecute(SlashCommandInteractionEvent event) {
+        user = event.getUser();
+        game = GameManager.getInstance().getUserActiveGame(user.getId());
+    }
+}
+
