+++ src/main/java/ti4/commands/uncategorized/AllInfo.java
@@ -1,6 +1,7 @@
 package ti4.commands.uncategorized;
 
 import java.util.List;
+
 import net.dv8tion.jda.api.entities.Member;
 import net.dv8tion.jda.api.entities.Role;
 import net.dv8tion.jda.api.events.interaction.command.SlashCommandInteractionEvent;
@@ -13,10 +14,10 @@ import ti4.commands.Command;
 import ti4.commands.cardsac.ACInfo;
 import ti4.commands.cardspn.PNInfo;
 import ti4.commands.cardsso.SOInfo;
-import ti4.commands.explore.RelicInfo;
 import ti4.commands.leaders.LeaderInfo;
 import ti4.commands.player.AbilityInfo;
 import ti4.commands.player.UnitInfo;
+import ti4.commands.relic.RelicInfo;
 import ti4.commands.tech.TechInfo;
 import ti4.generator.Mapper;
 import ti4.helpers.Constants;
