+++ src/main/java/ti4/helpers/AgendaHelper.java
@@ -43,10 +43,10 @@ import ti4.commands.cardsac.ACInfo;
 import ti4.commands.cardsac.DiscardACRandom;
 import ti4.commands.cardsac.DrawAC;
 import ti4.commands.cardsso.SOInfo;
-import ti4.commands.explore.DrawRelic;
 import ti4.commands.planet.PlanetExhaust;
 import ti4.commands.player.ClearDebt;
 import ti4.commands.player.SCPlay;
+import ti4.commands.relic.DrawRelic;
 import ti4.commands.special.RiseOfMessiah;
 import ti4.commands.special.SwordsToPlowsharesTGGain;
 import ti4.commands.special.WormholeResearchFor;
