+++ src/main/java/ti4/helpers/TransactionHelper.java
@@ -15,8 +15,8 @@ import net.dv8tion.jda.api.interactions.components.buttons.Button;
 import ti4.buttons.Buttons;
 import ti4.commands.cardsac.ACInfo;
 import ti4.commands.cardspn.PNInfo;
-import ti4.commands.explore.SendFragments;
 import ti4.commands.player.ClearDebt;
+import ti4.commands.relic.SendFragments;
 import ti4.commands.uncategorized.CardsInfo;
 import ti4.generator.Mapper;
 import ti4.helpers.Units.UnitType;
